refactor: renaming elasticsearch logger
